requirements.txt
pycaw 
pystray
pillow 
pywin32
comtypes